requirements.txt
pystemmer>=1.3.0
sklearn>=0.0
